requirements.txt
pydantic>=2.7
pydantic-settings>=2.4
requests>=2.31
jinja2>=3.1
python-dotenv>=1.0
tenacity>=8.2
openai>=1.35
anthropic>=0.34
